+++ example_application.properties
@@ -8,4 +8,3 @@ baseUrl=https://api.spotify.com/v1
 authorizationMethod=client_credentials
 #Setting none speeds up application start, servlet setup is mandatory for authorization_code method for getting user data
 spring.main.web-application-type=none
-#spring.main.allow-bean-definition-overriding=true
