+++ pom.xml
@@ -9,13 +9,15 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>pl.connectis</groupId>
-    <artifactId>Spotify_API_CLI</artifactId>
-    <version>0.1.0-ALPHA</version>
-    <name>Spotify_API_CLI</name>
+    <artifactId>spotify-api-cli</artifactId>
+    <version>0.1.1-ALPHA</version>
+    <name>Spotify API CLI</name>
     <description>Demo project for Spring Boot</description>
 
     <properties>
         <java.version>1.8</java.version>
+        <commons.cli.version>1.4</commons.cli.version>
+        <commons.io.version>2.6</commons.io.version>
     </properties>
 
     <dependencies>
@@ -35,12 +37,12 @@
         <dependency>
             <groupId>commons-cli</groupId>
             <artifactId>commons-cli</artifactId>
-            <version>1.4</version>
+            <version>${commons.cli.version}</version>
         </dependency>
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
-            <version>2.6</version>
+            <version>${commons.io.version}</version>
         </dependency>
