--- src/main/java/pl/connectis/Spotify_API_CLI/APICalls/IAPICaller.java
@@ -1,8 +1,0 @@
-package pl.connectis.Spotify_API_CLI.APICalls;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public interface IAPICaller {
-    void call(String ids);
-}
