--- src/main/java/pl/connectis/Spotify_API_CLI/DTOs/Artist.java
@@ -1,25 +1,0 @@
-package pl.connectis.Spotify_API_CLI.DTOs;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-import java.util.List;
-
-@Getter
-@Setter
-@ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Artist {
-    private ExternalURL external_urls;
-    private Followers followers;
-    private List<String> genres;
-    private String href;
-    private String id;
-    private List<Image> images;
-    private String name;
-    private Integer popularity;
-    private String type;
-    private String uri;
-}
