--- src/main/java/pl/connectis/Spotify_API_CLI/DTOs/Copyright.java
@@ -1,16 +1,0 @@
-package pl.connectis.Spotify_API_CLI.DTOs;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Copyright {
-
-    private String text;
-    private String type;
-}
