--- src/main/java/pl/connectis/Spotify_API_CLI/DTOs/Followers.java
@@ -1,16 +1,0 @@
-package pl.connectis.Spotify_API_CLI.DTOs;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Followers {
-
-    private String href;
-    private Integer total;
-}
