--- src/main/java/pl/connectis/Spotify_API_CLI/DTOs/Image.java
@@ -1,18 +1,0 @@
-package pl.connectis.Spotify_API_CLI.DTOs;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Image {
-
-    private Integer height;
-    private String url;
-    private Integer width;
-
-}
