--- src/main/java/pl/connectis/Spotify_API_CLI/authorization/AuthorizationCode.java
@@ -1,20 +1,0 @@
-package pl.connectis.Spotify_API_CLI.authorization;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class AuthorizationCode {
-
-    private String code;
-
-    public AuthorizationCode() {
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-}
