--- src/main/java/pl/connectis/Spotify_API_CLI/authorization/AuthorizationStrategy.java
@@ -1,7 +1,0 @@
-package pl.connectis.Spotify_API_CLI.authorization;
-
-import java.io.IOException;
-
-public interface AuthorizationStrategy {
-    void authorize() throws IOException;
-}
