+++ src/main/java/pl/connectis/spotifyapicli/APICalls/APICaller.java
@@ -1,0 +1,5 @@
+package pl.connectis.spotifyapicli.APICalls;
+
+public interface APICaller {
+    void call(String ids);
+}
