+++ src/main/java/pl/connectis/spotifyapicli/authorization/AuthorizationCode.java
@@ -1,0 +1,11 @@
+package pl.connectis.spotifyapicli.authorization;
+
+import lombok.Data;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+public class AuthorizationCode {
+
+    private String code;
+}
