+++ src/main/java/pl/connectis/spotifyapicli/authorization/AuthorizationCodeAuthorization.java
@@ -1,0 +1,119 @@
+package pl.connectis.spotifyapicli.authorization;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.ApplicationContext;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Component;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import java.io.IOException;
+import java.util.Base64;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.Condition;
+import java.util.concurrent.locks.ReentrantLock;
+
+@Slf4j
+@Component("authorization_code")
+public class AuthorizationCodeAuthorization implements AuthorizationStrategy {
+
+    private final AuthorizationCode authorizationCode;
+    private final ApplicationContext context;
+    private final RestTemplate restTemplate;
+    private final HttpHeaders httpHeaders;
+    private final ReentrantLock lock;
+    private final Condition condition;
+    private final AuthorizationConfig authorizationConfig;
+    @Value("{$spring.main.web-application-type}")
+    private String webApplicationType;
+
+
+    public AuthorizationCodeAuthorization(AuthorizationCode authorizationCode, ApplicationContext context, RestTemplate restTemplate, HttpHeaders httpHeaders, ReentrantLock lock, Condition condition, AuthorizationConfig authorizationConfig) {
+        this.authorizationCode = authorizationCode;
+        this.context = context;
+        this.restTemplate = restTemplate;
+        this.httpHeaders = httpHeaders;
+        this.lock = lock;
+        this.condition = condition;
+        this.authorizationConfig = authorizationConfig;
+    }
+
+    public void authorize() throws IOException {
+
+        if (!webApplicationType.equals("servlet")) {
+            throw new RuntimeException("Wrong spring web application type configuration for authorization_code method of authorization. Please change the method or set spring.main.web-application-type to servlet.");
+        }
+
+        HttpEntity<MultiValueMap<String, String>> httpEntity;
+
+        UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(authorizationConfig.getUserAuthorizationUri())
+                .queryParam("client_id", authorizationConfig.getClientID())
+                .queryParam("response_type", "code")
+                .queryParam("redirect_uri", "http://localhost:8080/collect");
+//				.queryParam("state", "")
+//				.queryParam("scope", "")
+//				.queryParam("show_dialog", "")
+
+        log.info(builder.build().toUri().toString());
+
+        try {
+            java.awt.Desktop.getDesktop().browse(builder.build().toUri());
+        } catch (IOException ex) {
+            log.error("Couldn't open default browser.");
+            ex.printStackTrace();
+        }
+
+        lock.lock();
+        try {
+            log.info("Awaiting 30 seconds for authorization code.");
+            condition.await(30, TimeUnit.SECONDS);
+        } catch (InterruptedException ex) {
+            log.info("Interrupted await for authorization code to be generated.");
+            ex.printStackTrace();
+            return;
+        } finally {
+            lock.unlock();
+        }
+
+        if (authorizationCode.getCode() == null) throw new IOException("Didn't get authorization code");
+
+        String toEncodeClientIdAndSecret = authorizationConfig.getClientID() + ":" + authorizationConfig.getClientSecret();
+        String encodedClientIdAndSecret = Base64.getEncoder().encodeToString(toEncodeClientIdAndSecret.getBytes());
+
+        httpHeaders.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
+        httpHeaders.set("Content-Type", "application/x-www-form-urlencoded");
+        httpHeaders.setBasicAuth(encodedClientIdAndSecret);
+        MultiValueMap<String, String> values = new LinkedMultiValueMap<>();
+        values.add("grant_type", authorizationConfig.getAuthorizationMethod());
+        values.add("code", authorizationCode.getCode());
+        values.add("redirect_uri", "http://localhost:8080/collect");
+        httpHeaders.putAll(values);
+
+        log.info(httpHeaders.toString());
+
+        httpEntity = new HttpEntity<>(values, httpHeaders);
+
+        Token newToken = restTemplate.postForObject(authorizationConfig.getAccessTokenUri(),
+                httpEntity,
+                Token.class);
+        log.info("ReceivedToken: {} ", newToken.toString());
+        newToken.saveToken();
+        Token tokenBeanToUpdate = context.getBean(Token.class);
+        updateTokenBean(tokenBeanToUpdate, newToken);
+    }
+
+    private void updateTokenBean(Token tokenBean, Token newToken) {
+        tokenBean.setAccessToken(newToken.getAccessToken());
+        tokenBean.setExpiresInSecond(newToken.getExpiresInSecond());
+        tokenBean.setExpirationTimeMillis(newToken.getExpirationTimeMillis());
+        tokenBean.setRefreshToken(newToken.getRefreshToken());
+        tokenBean.setScope(newToken.getScope());
+        tokenBean.setTokenType(newToken.getTokenType());
+        log.info("BeanToken: {}", tokenBean.toString());
+    }
+}
