+++ src/main/java/pl/connectis/spotifyapicli/authorization/AuthorizationStrategy.java
@@ -1,0 +1,7 @@
+package pl.connectis.spotifyapicli.authorization;
+
+import java.io.IOException;
+
+public interface AuthorizationStrategy {
+    void authorize() throws IOException;
+}
