+++ src/main/java/pl/connectis/spotifyapicli/dto/Artist.java
@@ -1,0 +1,24 @@
+package pl.connectis.spotifyapicli.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Artist {
+
+    @JsonProperty("external_urls")
+    private ExternalURL externalUrls;
+    private Followers followers;
+    private List<String> genres;
+    private String href;
+    private String id;
+    private List<Image> images;
+    private String name;
+    private Integer popularity;
+    private String type;
+    private String uri;
+}
