+++ src/main/java/pl/connectis/spotifyapicli/dto/Copyright.java
@@ -1,0 +1,12 @@
+package pl.connectis.spotifyapicli.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Copyright {
+
+    private String text;
+    private String type;
+}
