+++ src/main/java/pl/connectis/spotifyapicli/dto/ExternalID.java
@@ -1,0 +1,13 @@
+package pl.connectis.spotifyapicli.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ExternalID {
+
+    private String isrc;
+    private String ean;
+    private String upc;
+}
