+++ src/main/java/pl/connectis/spotifyapicli/dto/ExternalURL.java
@@ -1,0 +1,11 @@
+package pl.connectis.spotifyapicli.dto;
+
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ExternalURL {
+    private String spotify;
+}
