+++ src/main/java/pl/connectis/spotifyapicli/dto/Image.java
@@ -1,0 +1,14 @@
+package pl.connectis.spotifyapicli.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Image {
+
+    private Integer height;
+    private String url;
+    private Integer width;
+
+}
