+++ src/main/java/pl/connectis/spotifyapicli/dto/PagingObject.java
@@ -1,0 +1,16 @@
+package pl.connectis.spotifyapicli.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class PagingObject<T> {
+    private String href;
+    private T items;
+    private Integer limit;
+    private String next;
+    private Integer offset;
+    private String previous;
+    private Integer total;
+}
