+++ src/main/java/pl/connectis/spotifyapicli/dto/Restrictions.java
@@ -1,0 +1,10 @@
+package pl.connectis.spotifyapicli.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Restrictions {
+    private String reason;
+}
