--- src/test/java/pl/connectis/Spotify_API_CLI/SpotifyApiCliApplicationTests.java
@@ -1,15 +1,0 @@
-package pl.connectis.Spotify_API_CLI;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class SpotifyApiCliApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
