+++ src/test/java/pl/connectis/spotifyapicli/SpotifyApiCliApplicationTests.java
@@ -1,0 +1,15 @@
+package pl.connectis.spotifyapicli;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class SpotifyApiCliApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
